// packages/plugin-core/src/utils/md.d.ts
export type FoundRefT = {
    location: Location;
    matchText: string;
    isCandidate?: boolean | undefined;
    isFrontmatterTag?: boolean | undefined;
    note: NotePropsMeta;
};